refactor: create Directory and move to Attribute Component
